requirements.txt
mysql-connector-python==8.0.29
paho-mqtt==1.5.1
python-dotenv==0.19.2
